Rewrote initial stepsize, documentation changes.
The previous algorithm was trying to be very clever but did not do
well on "noisy" posteriors that came from simulated data (indirect
inference). This is hoped to be more robust.

At the same time, packaged parameters into a single structure, to make
further modifications easier.

Also started using DocStringExtensions.